+++ components/filemanager/api/src/main/java/com/flipperdevices/filemanager/api/share/ShareFile.kt
@@ -1,7 +1,0 @@
-package com.flipperdevices.filemanager.api.share
-
-data class ShareFile(
-    val name: String,
-    val flipperFilePath: String,
-    val size: Long
-)