+++ components/filemanager/impl/src/main/java/com/flipperdevices/filemanager/impl/model/ShareState.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.filemanager.impl.model
+
+data class ShareState(
+    val name: String,
+    val downloadProgress: DownloadProgress = DownloadProgress.Infinite(),
+    val processCompleted: Boolean = false
+)