+++ components/filemanager/sharecommon/src/main/java/com/flipperdevices/filemanager/sharecommon/model/ShareState.kt
@@ -1,8 +1,0 @@
-package com.flipperdevices.filemanager.sharecommon.model
-
-data class ShareState(
-    val name: String,
-    val downloadProgress: DownloadProgress = DownloadProgress.Infinite(),
-    val processCompleted: Boolean = false,
-    val dialogShown: Boolean = true
-)